fx: move Flicker
