+++ meson.build
@@ -96,6 +96,7 @@ sources = [
   'src/game/effect_routines/earthquake.c',
   'src/game/effect_routines/explosion.c',
   'src/game/effect_routines/finish_level.c',
+  'src/game/effect_routines/flicker.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/fmv.c',
@@ -152,7 +153,6 @@ sources = [
   'src/game/objects/effects/bubble.c',
   'src/game/objects/effects/earthquake.c',
   'src/game/objects/effects/explosion.c',
-  'src/game/objects/effects/flicker.c',
   'src/game/objects/effects/flipmap.c',
   'src/game/objects/effects/flood.c',
   'src/game/objects/effects/gunshot.c',
