+++ src/global/vars.c
@@ -4,7 +4,7 @@
 #include "game/effect_routines/earthquake.h"
 #include "game/effect_routines/explosion.h"
 #include "game/effect_routines/finish_level.h"
-#include "game/objects/effects/flicker.h"
+#include "game/effect_routines/flicker.h"
 #include "game/objects/effects/flipmap.h"
 #include "game/objects/effects/flood.h"
 #include "game/objects/effects/lara_effects.h"
@@ -22,7 +22,7 @@ void (*g_EffectRoutines[])(ITEM_INFO *item) = {
     Turn180,       FX_DinoStomp, LaraNormal,    FX_Bubbles,   FX_FinishLevel,
     FX_Earthquake, Flood,        RaisingBlock,  Stairs2Slope, DropSand,
     PowerUp,       FX_Explosion, LaraHandsFree, FxFlipMap,    LaraDrawRightGun,
-    FX_ChainBlock, Flicker,
+    FX_ChainBlock, FX_Flicker,
 };
 
 int16_t g_SampleLUT[MAX_SAMPLES] = { 0 };
